requirements.txt
asgiref==3.6.0
astroid==2.6.6
async-timeout==4.0.2
attrs==20.3.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.4
cryptography==36.0.1
Django==4.1.7
django-cors-headers==3.14.0
django-rest-passwordreset==1.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==3.9.2
gunicorn==20.1.0
isort==5.9.3
jellyfish==0.8.9
lazy-object-proxy==1.6.0
lib50==3.0.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
pexpect==4.8.0
platformdirs==2.5.1
platformshconfig==2.4.0
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
PyJWT==2.1.0
pylint==2.9.6
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rope==0.20.1
sqlparse==0.4.1
termcolor==1.1.0
toml==0.10.2
tomli==2.0.1
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.12.1
